# api/openapi.yaml
openapi: 3.1.0
info:
  title: Little Scout API
  description: |
    Little Scout — social restaurant ratings for parents and caregivers.
    Pilot metro: Dedham, Massachusetts (`pilot_city=dedham-ma`).
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Local Docker Compose
...
paths:
...
  /v1/restaurants:
    get:
      tags: [restaurants]
      summary: List restaurants in pilot city
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: Name search
        - name: cuisine
          in: query
          schema: { type: string }
          description: Filter by cuisine tag
      responses:
        "200":
          description: Restaurant list
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
            Cache-Control: { $ref: "#/components/headers/CacheControl" }
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestaurantSummary"
        "304":
          $ref: "#/components/responses/NotModified"
    post:
      tags: [restaurants]
      summary: Add restaurant (authenticated)
      security: [{ firebase: [] }]
      responses:
        "501":
          description: Not implemented — Firebase Auth required
  /v1/restaurants/map:
    get:
      tags: [restaurants]
      summary: Map entries with TTF/note/attribute counts
      description: |
        Returns restaurants with aggregated stats for the map view. An optional
        viewport bounding box narrows results; all four bounds must be supplied
        together (or none, for the whole pilot city).
      parameters:
        - name: min_lat
          in: query
          schema: { type: number, format: double, minimum: -90, maximum: 90 }
          description: South edge of viewport (requires all four bounds)
        - name: max_lat
          in: query
          schema: { type: number, format: double, minimum: -90, maximum: 90 }
          description: North edge of viewport (requires all four bounds)
        - name: min_lng
          in: query
          schema: { type: number, format: double, minimum: -180, maximum: 180 }
          description: West edge of viewport (requires all four bounds)
        - name: max_lng
          in: query
          schema: { type: number, format: double, minimum: -180, maximum: 180 }
          description: East edge of viewport (requires all four bounds)
      responses:
        "200":
          description: Map entries
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
            Cache-Control: { $ref: "#/components/headers/CacheControl" }
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestaurantMapEntry"
        "304":
          $ref: "#/components/responses/NotModified"
        "422":
          description: Invalid bbox (partial bounds or min not less than max)
  /v1/restaurants/seed-jobs:
    post:
      tags: [restaurants]
      summary: Start a background restaurant seed job for a ZIP, city, or coordinates
      security: [{ firebase: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantSeedJobRequest"
      responses:
        "202":
          description: Seed job accepted or reused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantSeedJobResponse"
        "400":
          description: Invalid or un-geocodable location
        "503":
          description: Server-side Maps key is not configured
  /v1/restaurants/seed-jobs/{job_id}:
    get:
      tags: [restaurants]
      summary: Poll a restaurant seed job
      security: [{ firebase: [] }]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Seed job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantSeedJobResponse"
        "404":
          description: Seed job not found
  /v1/restaurants/{id}:
    get:
      tags: [restaurants]
      summary: Restaurant detail with TTF aggregates
      parameters:
        - $ref: "#/components/parameters/RestaurantId"
      responses:
        "200":
          description: Restaurant detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantDetailResponse"
        "404":
          description: Not found
  /v1/restaurants/{id}/ttf:
    get:
      tags: [restaurants]
      summary: Aggregated TTF stats
      parameters:
        - $ref: "#/components/parameters/RestaurantId"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TtfAggregate"
    post:
      tags: [restaurants]
      summary: Submit TTF observation (authenticated)
      security: [{ firebase: [] }]
      parameters:
        - $ref: "#/components/parameters/RestaurantId"
      responses:
        "501":
          description: Not implemented — Firebase Auth required
  /v1/restaurants/{id}/attributes:
    get:
      tags: [restaurants]
      summary: Aggregated attribute ratings
      parameters:
        - $ref: "#/components/parameters/RestaurantId"
      responses:
        "200":
          description: Attribute aggregates (stub)
    post:
      tags: [restaurants]
      summary: Submit attribute rating (authenticated)
      security: [{ firebase: [] }]
      parameters:
        - $ref: "#/components/parameters/RestaurantId"
      responses:
        "501":
          description: Not implemented — Firebase Auth required
  /v1/restaurants/{id}/notes:
    get:
      tags: [restaurants]
      summary: List restaurant notes
      parameters:
        - $ref: "#/components/parameters/RestaurantId"
      responses:
        "200":
          description: Notes list
    post:
      tags: [restaurants]
      summary: Add note (authenticated)
      security: [{ firebase: [] }]
      parameters:
        - $ref: "#/components/parameters/RestaurantId"
      responses:
        "501":
          description: Not implemented — Firebase Auth required
  /v1/restaurants/search:
    get:
      tags: [restaurants]
      summary: Radius query — restaurants near a point, sorted by distance
      description: |
        Public (no auth). Returns `RestaurantMapEntry` rows within `radius_m` of
        the given coordinates, ordered by Haversine distance ascending.
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number, format: double }
          description: Center latitude
        - name: lng
          in: query
          required: true
          schema: { type: number, format: double }
          description: Center longitude
        - name: radius_m
          in: query
          schema: { type: integer, default: 8000, minimum: 500, maximum: 25000 }
          description: Search radius in metres (clamped 500–25000)
        - name: q
          in: query
          schema: { type: string }
          description: Optional name filter (ILIKE)
        - name: limit
          in: query
          schema: { type: integer, default: 100, minimum: 1, maximum: 250 }
          description: Max results (clamped 1–250)
      responses:
        "200":
          description: Distance-sorted restaurant list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestaurantMapEntry"
...
components:
  securitySchemes:
    firebase:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Auth JWT (Apple Sign-In) — Phase 3
  parameters:
    RestaurantId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  headers:
    ETag:
      description: Strong validator for conditional requests (If-None-Match).
      schema: { type: string }
      example: '"3a7bd3e2360a3d2a..."'
    CacheControl:
      description: Caching policy for clients and shared caches.
      schema: { type: string }
      example: "public, max-age=30, stale-while-revalidate=300"
  responses:
    NotModified:
      description: |
        Cached copy is still fresh. Returned when If-None-Match matches the
        current ETag. Empty body; ETag and Cache-Control are repeated.
      headers:
        ETag: { $ref: "#/components/headers/ETag" }
        Cache-Control: { $ref: "#/components/headers/CacheControl" }
  schemas:
...
    RestaurantSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        address: { type: string }
        lat: { type: number, format: double }
        lng: { type: number, format: double }
        cuisine_tags: { type: array, items: { type: string } }
        pilot_city: { type: string }
        google_place_id: { type: string, nullable: true }
        google_maps_url: { type: string, nullable: true }
    RestaurantDetail:
      allOf:
        - $ref: "#/components/schemas/RestaurantSummary"
        - type: object
          properties:
            google_place_id: { type: string, nullable: true }
            google_maps_url: { type: string, nullable: true }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    TtfAggregate:
      type: object
      properties:
        sample_size: { type: integer }
        median_minutes: { type: number, nullable: true }
        avg_quality: { type: number, nullable: true }
        last_updated: { type: string, format: date-time, nullable: true }
    RestaurantDetailResponse:
      type: object
      properties:
        restaurant: { $ref: "#/components/schemas/RestaurantDetail" }
        ttf: { $ref: "#/components/schemas/TtfAggregate" }
...
    RestaurantSeedJobRequest:
      type: object
      properties:
        location:
          type: string
          nullable: true
          description: ZIP, city/state, or address to geocode before seeding
          example: "02026"
        lat:
          type: number
          format: double
          nullable: true
        lng:
          type: number
          format: double
          nullable: true
        radius_m:
          type: integer
          default: 8000
          minimum: 1000
          maximum: 25000
        force:
          type: boolean
          default: false
      oneOf:
        - required: [location]
        - required: [lat, lng]
    RestaurantSeedJob:
      type: object
      properties:
        id: { type: string, format: uuid }
        pilot_city: { type: string }
        area_key: { type: string }
        query: { type: string, nullable: true }
        lat: { type: number, format: double }
        lng: { type: number, format: double }
        radius_m: { type: integer }
        refresh: { type: boolean }
        status:
          type: string
          enum: [pending, running, succeeded, failed, skipped]
        error: { type: string, nullable: true }
        inserted_count: { type: integer }
        updated_count: { type: integer }
        closed_count: { type: integer }
        outside_area_count: { type: integer }
        skipped_count: { type: integer }
        out_of_area_count: { type: integer }
        unique_places_count: { type: integer }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    RestaurantSeedJobResponse:
      type: object
      properties:
        job: { $ref: "#/components/schemas/RestaurantSeedJob" }
        reused: { type: boolean }
    RestaurantMapEntry:
      allOf:
        - $ref: "#/components/schemas/RestaurantSummary"
        - type: object
          properties:
            ttf:
              $ref: "#/components/schemas/TtfAggregate"
            note_count: { type: integer }
            attribute_rating_count: { type: integer }
            watched: { type: boolean, default: false }
...